sql_syntax_validator/requirements.txt
click
sqlparse
psycopg2-binary
